requirements.txt
torch
transformers
datasets
numpy
tqdm
google-generativeai
python-dotenv
